chore(team): split step lifecycle helpers
