+++ src/team/manager/step_materialization.rs
@@ -2,7 +2,7 @@ use serde_json::Value;
 use sqlx::Sqlite;
 use uuid::Uuid;
 
-use super::step_lifecycle::load_step_record_tx;
+use super::step_helpers::load_step_record_tx;
 use super::{
     TeamManager, TeamRunEventRecord, TeamStepStatus, TeamTaskRecord, TeamTaskStepExecutionSpec,
     parse_task_execution_plan, team_step_status_to_str, validate_task_execution_steps,
